+++ projects/ngx-dashboard/src/lib/providers/index.ts
@@ -1,1 +1,3 @@
 export * from './cell-settings-dialog';
+export * from './empty-cell-context';
+export * from './unknown-widget';